+++ GloryECS/GloryReflectStatic/TypeData.h
@@ -18,7 +18,7 @@ namespace GloryReflect
 
 	private:
 		const char* m_TypeName;
-		const size_t m_TypeHash;
+		const uint32_t m_TypeHash;
 		const int m_FieldCount;
 		const bool m_BasicType;
 		const FieldData* m_pFields;
